Added Season.CURRENT
I got the idea to add a Season.CURRENT constant, which always refers to
the current season. This means, the value of this key will update for
Preseason2016 and Season2016 as soon as they arrive. This way
applications don't have to care about updating their constants, if they
only want fresh data.

Also cleaned up the other enums.