Removed ItemListData and MasteryListData
The methods which used those Objects, now use ItemData and MasteryData
respectively. The enums were exactly identical in both cases. Also we
didn't have duplicate ChampListData and SpellListData objects, so it
should be appropriate and consistent to remove those 2 classes

(We still distinguish between RuneData and RuneListData, because those
enums are not exactly identical)