Fix bug with getSummonerByName
The keys for the map of summoners are the summoner name in lowercase
with no white space, so that's what we need to search for to see if a
summoner is present. Now you can call getSummonerByName("Rithms"),
getSummonerByName("R   i  T  hm       S"), getSummonerByName("r    i
thms"), etc. and it won't give a 404 response.